refactor(ConfettiCanvas): extract into reusable common component (v2)
